+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0-beta8
+**Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
+**Date**: Fri May 10 09:52:03 CEST 2019
+
+* CORE: [CHANGE] If --debug was applied and an exception occured during login operation the exception gets dumped to stdout
+
+
 ## 1.0.0-beta7
 **Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
 **Date**: Fri May 10 09:52:03 CEST 2019